Set the CAPROVER_GIT_COMMIT_SHA env var automatically when building
When we are able to and the environment variable isn't already in use,
we set the CAPROVER_GIT_COMMIT_SHA env var to the git has that is being
deployed. This allows it to be used when building the image and can be
copied into the image to use a runtime as well.

Fixes #922